+++ CHANGELOG.md
@@ -4,6 +4,11 @@
 
 ## [Unreleased]
 
+## [1.0.10] - 2025-11-17
+
+- Fix bug with time tracking after project closure
+- Update plugin description
+
 ## [1.0.9] - 2025-11-16
 
 - Add status bar widget with daily coding activity
@@ -45,7 +50,8 @@
 
 - Support IntelliJ Platform 2024.3.5
 
-[Unreleased]: https://github.com/codeclocker/codeclocker-intellij-plugin/compare/v1.0.9...HEAD
+[Unreleased]: https://github.com/codeclocker/codeclocker-intellij-plugin/compare/v1.0.10...HEAD
+[1.0.10]: https://github.com/codeclocker/codeclocker-intellij-plugin/compare/v1.0.9...v1.0.10
 [1.0.9]: https://github.com/codeclocker/codeclocker-intellij-plugin/compare/v1.0.8...v1.0.9
 [1.0.8]: https://github.com/codeclocker/codeclocker-intellij-plugin/compare/v1.0.7...v1.0.8
 [1.0.7]: https://github.com/codeclocker/codeclocker-intellij-plugin/compare/v1.0.6...v1.0.7